error edited
The error in the code stems from misusing the `super()` function within the `__init__()` method of the `Net` class. The error message, "AttributeError: type object 'Net' has no attribute 'self'", indicates that the `super()` function is being incorrectly invoked. In Python, when using `super()` to call the parent class's `__init__()` method, it should be written as `super(ChildClass, self).__init__()`, where `ChildClass` refers to the subclass inheriting from the parent class. However, in the provided code, `Net.self` is used instead, causing the AttributeError. This error prevents the initialization of the `Net` class properly. To fix it, the `__init__()` method of the `Net` class should be corrected to use `super()` function appropriately, ensuring that the subclass inherits the behavior of the parent class correctly.